chore: refactor of notebook creation
